+++ CHANGES.md
@@ -19,6 +19,8 @@
     -   Currently uses only a case-sensitive check, using 'in' (i.e. if the filtering parameter is inside the job-name, then it passes the check).
 -   Added `--sort` argument to sort 'sacct' results by the given field.
     -   New method `sacct._sort_lines()` does this, called from `sacct.sacct_results()`.
+-   Added `--id` argument to filter by 'JobID' number.  Also works for canceling jobs.
+    -   New method `sacct._filter_by_jobid()` does the filtering by parsing the command-line argument specification string.  This is called by `sacct._filter_lines()`.
 -   By default, reformat the times-strings output by sacct (replace 'T' separator with optional alternative character.  Set by parameters in `const.py`, with code in `sacct._parse_sacct_line()`.
 -   By default, printed output now only shows a subset of available parameters.  Use '--verbose' to include all of them.
     -   New method `utils._filter_fields_in_lines()` is called from within `utils.print_lines_dicts()`.